FAB-13372 Fabric-Samples return error msg
Update the Balance Transfer to return the fabric error.
The sample application was updated to both return the error
in the REST response and to post the error to the log.
The sample was updated with the 'channel.close()' so
that users will note how to shutdown a channel and not
hold connections open.

Change-Id: I49f20a50340adff52cf57db00a121ffd75eb1827
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>